PABLO: add a function to find a Morton in a sorted list of octants
Given a target Morton number and a sorted list of octants, finds the
index of the last octant whose Morton number if lower or equal to the
target Morton number. If the target Morton is lower than the Morton
of the first element, the index of the past-the-element element is
returned.